speedup test_smallbaselineApp.py (#147)
* speedup test_smallbaselineApp.py

add 3 template files into local `test` directory and turn OFF pixelwise inversion for both invert_network and correct_topography step to speed up the testing process.

add --test-pyaps option and keep existing ECMWF file

* move unittestGalapagos* to test dir

move unittestGalapagosSenDT128.template as a more appropriate location

remove files in mintpy.troposphericDelay.weatherModel = ECMWF from docs/examples/input_files because 1) they are not recommended for best performance and 2) they are copies of them in test dir now with fast processing setup.

+++ test/test_smallbaselineApp.py
@@ -9,35 +9,45 @@ import sys
 import argparse
 import subprocess
 import numpy as np
+from mintpy import smallbaselineApp
 
 
-# example datasets
-pDict = {
-    'KujuAlosAT422F650' : 'https://zenodo.org/record/2748170/files/KujuAlosAT422F650.tar.xz',
-    'WellsEnvD2T399'    : 'https://zenodo.org/record/2748560/files/WellsEnvD2T399.tar.xz',
-    'FernandinaSenDT128': 'https://zenodo.org/record/2748487/files/FernandinaSenDT128.tar.xz',
-}
-pNames = list(pDict.keys())
+URL_LIST = [
+    'https://zenodo.org/record/2748170/files/KujuAlosAT422F650.tar.xz',
+    'https://zenodo.org/record/2748560/files/WellsEnvD2T399.tar.xz',
+    'https://zenodo.org/record/2748487/files/FernandinaSenDT128.tar.xz',
+]
+
+PROJ_NAME_LIST = [os.path.basename(url).split('.')[0] for url in URL_LIST]
+TEMPLATE_FILE_LIST = [os.path.join(os.path.dirname(__file__), '{}.txt'.format(proj_name))
+                      for proj_name in PROJ_NAME_LIST]
 
 
 #####################################################################################
 EXAMPLE = """example:
   $MINTPY_HOME/test/test_smallbaselineApp.py
-  $MINTPY_HOME/test/test_smallbaselineApp.py  KujuAlosAT422F650
+  $MINTPY_HOME/test/test_smallbaselineApp.py  --dset KujuAlosAT422F650
   $MINTPY_HOME/test/test_smallbaselineApp.py  --nofresh
-"""
+
+  # available test datasets:
+  {}
+""".format(PROJ_NAME_LIST)
 
 def create_parser():
     parser = argparse.ArgumentParser(description='Test smallbaselineApp workflow with example datasets.',
                                      formatter_class=argparse.RawTextHelpFormatter,
                                      epilog=EXAMPLE)
 
-    parser.add_argument('--dset', dest='dset_name', choices=pNames+['all'], default='all',
+    parser.add_argument('--dset', dest='dset_name', choices=PROJ_NAME_LIST+['all'], default='all',
                         help='name(s) of datasets to be tested.\n'+
-                             'Available datasets: {}\n'.format(pNames)+
+                             'Available datasets: {}\n'.format(PROJ_NAME_LIST)+
                              'Default is "all" for ALL DATASETS.')
+    parser.add_argument('--test-pyaps', dest='test_pyaps', action='store_true',
+                        help='Include testing of PyAPS.')
+
     parser.add_argument('--dir', dest='test_dir', default='~/insarlab/test',
                         help='test directory. Default: ~/insarlab/test')
+
     parser.add_argument('--nofresh', dest='fresh_start', action='store_false',
                         help='Use exsiting files WITHOUT starting from the scratch.')
     return parser
@@ -46,32 +56,40 @@ def create_parser():
 def cmd_line_parse(iargs=None):
     parser = create_parser()
     inps = parser.parse_args(args=iargs)
+
     inps.test_dir = os.path.expanduser(inps.test_dir)
 
     if inps.dset_name.lower() == 'all':
-        inps.dset_name = pNames
+        inps.dset_name = PROJ_NAME_LIST
+
     if isinstance(inps.dset_name, str):
         inps.dset_name = [inps.dset_name]
     return inps
 
 
 #####################################################################################
-def run_dataset(dset_name, test_dir, fresh_start=True):
+
+
+def test_dataset(dset_name, test_dir, fresh_start=True, test_pyaps=False):
     print('Go to test directory:', test_dir)
     os.chdir(test_dir)
 
+    dset_idx = PROJ_NAME_LIST.index(dset_name)
+    dset_url = URL_LIST[dset_idx]
+    template_file = TEMPLATE_FILE_LIST[dset_idx]
+
     # download tar file
-    tar_file = '{}.tar.xz'.format(dset_name)
+    tar_file = os.path.basename(dset_url)
     if not os.path.isfile(tar_file):
         print('downloading tar file ...')
-        cmd = 'wget {}'.format(pDict[dset_name])
+        cmd = 'wget {}'.format(dset_url)
         print(cmd)
         os.system(cmd)
-    print('tar file exists, skip downloading.')
+    print('tar file exists, skip re-downloading.')
 
     # uncompress tar file
     if not fresh_start and os.path.isdir(dset_name):
-        print('use existing files in {}'.format(dset_name))
+        print('use existing files in project directory: {}'.format(dset_name))
     else:
         # remove existing directory
         if os.path.isdir(dset_name):
@@ -85,20 +103,19 @@ def run_dataset(dset_name, test_dir, fresh_start=True):
         print(cmd)
         os.system(cmd)
 
-    # runing smallbaselineApp
+    # set working directory
     work_dir = os.path.join(test_dir, dset_name, 'mintpy')
     os.chdir(work_dir)
     print('Go to work directory:', work_dir)
 
-    cmd = 'rm ./inputs/ECMWF.h5'
-    os.system(cmd)
-    print('remove existing tropospheric delay file: ./inputs/ECMWF.h5')
+    # remove pyaps existing products or not
+    if test_pyaps:
+        cmd = 'rm ./inputs/ECMWF.h5'
+        os.system(cmd)
+        print('remove existing tropospheric delay file: ./inputs/ECMWF.h5')
 
-    cmd = 'smallbaselineApp.py $MINTPY_HOME/docs/examples/input_files/{}.txt'.format(dset_name)
-    cmd = os.path.expandvars(cmd)
-    status = subprocess.Popen(cmd, shell=True).wait()
-    if status is not 0:
-        raise RuntimeError('Test failed for example dataset {}'.format(dset_name))
+    # runing smallbaselineApp
+    smallbaselineApp.main([template_file])
     return
 
 
@@ -111,14 +128,15 @@ def main(iargs=None):
         dset_name = inps.dset_name[i]
         print('-'*50)
         print('Start testing smallbaselineApp workflow on exmaple dataset {}/{}: {}'.format(i+1, num_dset, dset_name))
-        run_dataset(dset_name,
-                    test_dir=inps.test_dir,
-                    fresh_start=inps.fresh_start)
-        print('Pass testing smallbaselineApp workflow on exmaple dataset {}/{}: {}'.format(i+1, num_dset, dset_name))
+        test_dataset(dset_name,
+                     test_dir=inps.test_dir,
+                     fresh_start=inps.fresh_start,
+                     test_pyaps=inps.test_pyaps)
+        print('PASS testing smallbaselineApp workflow on exmaple dataset {}/{}: {}'.format(i+1, num_dset, dset_name))
 
-    if num_dset == len(pNames):
+    if num_dset == len(PROJ_NAME_LIST):
         print('-'*50)
-        print('Pass ALL testings without runing errors.')
+        print('PASS ALL testings without running errors.')
         print('-'*50)
     return
